Create models for Profiles, Posts and Votes
To build our social network application, we need to store data. We
store data in a database using Django Models. These models are defined
in models.py
Here we create 3 models:
Profile - stores data about a User such as their username and picture
Post - an individual post on our social network. Posts are simply a
string of text associated with a Profile
Vote - an upvote on a Post: any Profile can add a Vote associated with
any Post