Create views for posts
In the last task, we added a serializer to convert Posts to and from
our API representations.

After adding the code in this task, use your favorite REST interface
(such as
Postman) to try some requests on http://localhost:8000/api/posts - at
this point we've implemented support for listing all posts (GET to
/api/posts), creating a post (POST to /api/posts), getting an
individual post by ID (GET to /api/posts/<id>), updating a post (PUT to
/api/posts/<id>) and deleting a post (DELETE to /api/posts/<id>)

As you might have noticed, that's a lot of code to do not a whole lot.
Fortunately Django REST Framework provides a lot of conveniences for us
that we'll explore in later tasks.